[T11] UserContext moved to rest module
