+++ core/src/main/java/com/tinqinacademy/bffservice/core/converters/hotel/bookroom/BookRoomBffInputToBookRoomInput.java
@@ -2,18 +2,15 @@ package com.tinqinacademy.bffservice.core.converters.hotel.bookroom;
 
 import com.tinqinacademy.bffservice.api.operations.hotelservice.hotel.bookroom.BookRoomBffInput;
 import com.tinqinacademy.bffservice.core.converters.BaseConverter;
-import com.tinqinacademy.bffservice.persistence.model.context.UserContext;
 import com.tinqinacademy.hotel.api.operations.hotel.bookroom.BookRoomInput;
 import org.springframework.stereotype.Component;
 
 @Component
 public class BookRoomBffInputToBookRoomInput extends BaseConverter<BookRoomBffInput, BookRoomInput, BookRoomBffInputToBookRoomInput> {
 
-    private final UserContext userContext;
 
-    public BookRoomBffInputToBookRoomInput(UserContext userContext) {
+    public BookRoomBffInputToBookRoomInput() {
         super(BookRoomBffInputToBookRoomInput.class);
-        this.userContext = userContext;
     }
 
     @Override
@@ -22,7 +19,7 @@ public class BookRoomBffInputToBookRoomInput extends BaseConverter<BookRoomBffIn
         BookRoomInput input = BookRoomInput.builder()
                 .startDate(bffInput.getStartDate())
                 .endDate(bffInput.getEndDate())
-                .userId(userContext.getUserId())
+                .userId(bffInput.getUserContextId())
                 .build();
 
         return input;
