+++ rest/src/main/java/com/tinqinacademy/bffservice/rest/security/JwtAuthFilter.java
@@ -2,7 +2,6 @@ package com.tinqinacademy.bffservice.rest.security;
 
 import com.tinqinacademy.authenticationservice.api.operations.auth.AuthOutput;
 import com.tinqinacademy.authenticationservice.restexport.AuthenticationRestExport;
-import com.tinqinacademy.bffservice.persistence.model.context.UserContext;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -56,7 +55,7 @@ public class JwtAuthFilter extends OncePerRequestFilter {
                 .build();
         Authentication authentication = new CustomAuthToken(userDetailsModel);
         SecurityContextHolder.getContext().setAuthentication(authentication);
-        userContext.setUserId(userId);
+        userContext.setContext(userId);
 
         filterChain.doFilter(request, response);
     }
